Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

Thank you for this contribution.  As it turns out, there is a more complete implementation of this fix already underway upstream, which handles additional Intent data, mailto: query parameters, and includes a modest set of unit tests.  I encourage you to take a look at it when it is merged into the public sources.  

I appreciate your work on this and I look forward to future patches.

Patch-set: 1