When replying in Email, do not set message id on new message.
When replying in Email application, the application deliberately
set the message id of the new message as message id of the original
message. The reason for this was to identify multiple draft messages.
However, I have found out this is not necessary. Moreover, when sending
a new message with the same message id, many mail stores do not accept the
new message at all, so this is a quite serious bug.

This has been reported as issue #3030.