Update patch set 2
Patch Set 2: No score; No score

Patrick-

This is really great and I appreciate your quick work on preparing this patch.  Most of the comments below are simple "style guide" cleanups.  The only code issue that I would like to see changed is to make the unit tests more symmetric;  You should be applying the same tests to the two charsets (more like a matrix).

I have one other request which is to ensure that the master comment for the CL wraps at 80 columns in lines 3 and beyond (this is a git requirement.)

Thanks!
Andy

Patch-set: 2
Label: Code-Review=0
Label: Verified=0