Update patch set 3
Patch Set 3: Verified; No score

Verified - tests working.

There's only one thing left to fix, which is the master comment (for the patch).  You need to make sure that it wraps at 80 columns in lines 3 and beyond (this is a git requirement.)  With that fixed we can +2 and submit.

Thanks! Andy

e.g. something like:
This change allows emails with a
  "Content-Type: text/plain ; charset=iso-8859-1"
header to be parsed as plain text.
Added unit tests for whitespace and the charset parameter.

Patch-set: 3
Label: Code-Review=0
Label: Verified=+1