Update patch set 2
Patch Set 2: No score

JBQ is correct, the code looks great (thanks!) but unit tests are required.  Feel like writing some?

Your baseline should be to execute this command:
  runtest -c com.android.email.mail.transport.SmtpSenderUnitTests email

And have it run the existing tests to completion.

The challenge here - which won't be terrible - is that the current version of MockTransport doesn't support reopenTls().  You'll need to implement that;  It doesn't really need to do much except possibly reset state, and set a new "mTlsReopened" flag.  There should be a getter as well, e.g. boolean getTlsReopened() so the tests can confirm that it happened.

Then, you could fairly easily clone SmtpSenderUnitTests.testSimpleLogin() to a pair of TLS variants, with the "working" and "failing" response ordering, and make sure that they both pass with your fix.

Patch-set: 2
Label: Verified=0