Update patch set 2
Patch Set 2: Fails

This change has some large conflicts on the latest master. It might be relevant to consider abandoning this change if the functionality that it brought in was fixed independently by another change (which could explain the conflicts).

Patch-set: 2
Reviewer: Jean-Baptiste Queru <1000660@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=-1