Update patch set 2
Patch Set 2: No score; I would prefer that you didn't submit this

I need to see this in the context of some actual use cases.  Is there a bug report with some repro steps?  That would help a lot, and if so, it should be in the change comments for reference.

The underlying parser shouldn't be failing if you are entering a valid address.  Also, the underlying parser has changed since this patch was submitted.  So I'm not ready to accept this until I understand the fix a little better.

Finally, as JBQ noted, unit tests are needed here.  They would capture both the problem and the fix, as well as safeguard against future regressions.  And the good news is that it would be very easy to test this class, as it stands alone and the test can be very lightweight.

Patch-set: 2
Reviewer: Andy Stadler <1001252@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0