Update patch set 2
Patch Set 2: No score

I will look into writing unit tests, please standby.

Yes, this bug has been reported as issues 1393, 1950 and 2920.

It is not the email address that is the problem, but the real name of the recipient, i.e. "John Doé". If the name before encoding it contains a 8 bit character, it is discarded. Note this is before encoding it, after encoding it with for example Quoted printable or Base64, it must of course not contain any 8 bit characters.

Patch-set: 2
Reviewer: Marek Pola <1002263@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0