Update patch set 2
Patch Set 2:

Marek, what software version are you running?  Now that I'm looking at the bugs, this is something that was fixed in the "donut" software release.

(Take a look in AddressUnitTests.java - you should find testParsingUtf16Name and testParsingUtf32Name - which were added to cover this type of bug.)

Note - unit tests would be very helpful and are still welcome! But you may find that on a donut or eclair version of the software, they already pass?

Patch-set: 2