Update patch set 1
Patch Set 1:

Okay, I discussed this with Marc earlier this year and, if I remember correctly, he said that was the main reason he didn't want to try to implement this.

If you make it work with large messages by rendering the partial MIME message or fall back to the current behavior of plain text, it would be more likely to be accepted.

Patch-set: 1