Update patch set 1
Patch Set 1:

Hi Mirko-

I'm referring to the bug you wrote up, http://code.google.com/p/android/issues/detail?id=5078

Re this change, when I asked you to add the bug reference to the change, it actually needs to be in the comment (you need to edit the git commit message and re-upload it.)  When a given change is accepted into the tree, only the change itself (with its comment) is what's recorded permanently - not the comment stream here in Gerrit.  The reference back to the bug being fixed is an important part of the patch.

Re my comments in the bug, I think that bug itself needs to be more fully diagnosed before we start making changes in the code, so we should probably continue discussion in that venue before we return to updating/validating this patch.

Thanks
Andy

Patch-set: 1