Update patch set 1
Patch Set 1:

Just to be clear, I fully understood your request to edit the commit log, but it seems to me there's no point in doing that if the bug needs to be discussed first. Or is it required in any case for bureacratic reasons?

And, I really hope to see a comment in the issue tracker explaining why this change "shouldn't be needed". You picked my curiosity now.

Patch-set: 1