Update patch set 2
Patch Set 2:

At this point it's likely that any change submitted in this area will have severe merge conflicts with Google's internal tree (which has been evolving independently for many months, and working on it immediately might not be an efficient use of time.

I'll keep this change around until that code gets open-sourced, so that it's possible to better evaluate how to proceed.

Patch-set: 2