Update patch set 1
Patch Set 1:

Chris,

I've got great news, and bad news.

First, the great news: this fixes an important issue indeed, and we thank you for helping Android get better.

Now, the bad news: this turned out to be important enough that we've fixed it internally as well. We haven't had a chance to open-source our fix yet.

Here's my suggestion: I think the best solution is going to be to accept your change into the AOSP tree, but to merge it as a no-op to Google's internal master tree (since we already have a functionally equivalent change there). This way, a record of your contribution will exist in the permanent history so that you get properly credited, and your fix will be in the AOSP tree until we reach a point where our version of that fix gets open-sourced there.

I hope that this won't discourage you from contributing other changes to AOSP. You might find in the future that you can discuss your contribution plans on the android-contrib, which would hopefully give you more advance notice about such situations. Also, if you happen to come aware of changes by other authors that would benefit AOSP, it'd be great if you could encourage them to also contribute those changes to AOSP: that'd benefit millions of Android users and would be one less change that someone would need to rebase across multiple versions of Android over time.

Once again, thank you for your involvement in the Open-Source community around Android.

Patch-set: 1