Update patch set 1
Patch Set 1:

Chris -

As JBQ noted, thanks very much for this contribution, and I apologize for the fact that it was a duplicate of fixes that were still "in the pipeline".

If you wish to push forward with this contribution (and I agree with JBQ, you should), then I'd like to propose that you revise it slightly.  You see, I'm typically fairly strict about requiring unit tests to go along with patches, and unfortunately, the change you made is in code that is difficult to unit test (because it's busy talking to a system service).

However, there's a simple alternative.  Instead of catching this condition in isActive(), you can write similar code in the PolicySet constructor, along the lines of:

            if (passwordMode == PASSWORD_MODE_NONE) {
                maxPasswordFails = 0;
                maxScreenLockTime = 0;
                minPasswordLength = 0;
            }

This generalizes your fix to cover some additional edge cases;  And, you can now write unit tests for these cases (see SecurityPolicyTests.java, which already includes a unit test for this constructor).

I hope this is helpful, and I look forward to reviewing the change.

Patch-set: 1
Reviewer: Andy Stadler <1001252@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0