Update patch set 10
Patch Set 10:

(Apologies for the very long delay).

This change builds and merges fine in the open-source master tree. At a glance, the code style looks fine to me.

However, there's a conflict when trying to merge it to the internal Google master.

Andy, it'd be good if you could have a look at the conflict ahead of time. If that conflict looks reasonable, feel free to mark this change as verified, and to submit it when our internal merges look up-to-date and when you've got time to deal with the conflict resolution.

Patch-set: 10