Update patch set 1
Patch Set 1: Looks good to me, but someone else must approve

I originally submitted this (incorrectly) to the Froyo branch.

At that time it was mentioned that I may need to supply unit tests for this change.

While I understand the essential idea of unit tests, I have not used them to this point.  For a change like this, would it involve injecting a synthetic volume up/down to this activity to test the function of these two methods?

Please advise if I would have to supply this test and if I have the correct idea for the test. (I am 6 months into programming Android and in general - most of my testing involves beating up my day-to-day phone with my code!)

Thanks

Patch-set: 1