Create change
Patch-set: 1
Change-id: I7581123f21fbb4015153ca6f4a0c14c0f6a769fc
Subject: A email apps uses a context menu served by google. But sometimes a parameter "MenuItem" in the "onContextItemSelected" was passed wrongly. So switched to global variable and initiate it in the "onContextItemSelected" because a parameter "MenuItem" in the "onCreateContextMenu" is always correct. And global variable is used instead of a parameter while forwarding.
Branch: refs/heads/gingerbread
Commit: 5dc9b433311bc65ef671766c533fa1cb4c739005