Update patch set 2
Patch Set 2: (1 inline comment)

I think you are fixing a bug we have also discovered, which is that android.util.Base64InputStream will throw an IOException on bad base64 data, but the MIME implementation simply skips over non-base64-characters.  We are fixing this in honeycomb by throwing a new, different exception on bad base64 data.

For gingerbread, this change is OK, but we're not going to merge it into Honeycomb.

Meanwhile, we need to make sure that we're properly handling the errors that you are seeing.  I would like you to add a unit test that calls decodeBody() twice, once with good base64 data, and once with bad base64 data.  The unit test should fail on the old code (because it throws IOException) and succeed on the new code (no exception, and at least partial decoding of the input.)  There is already a test framework for this class, see MimeUtilityTest.java

Patch-set: 2
Label: Code-Review=0
Label: Verified=0