Minimal support for Exchange ActiveSync 12.1
Main goal was to allow support of ActiveSync 12.1 and keep 100% backward
compatible. The two additional tags in "searchGal" can be omitted from this
patch (they will give better results, but do carry a performance penalty)...

This was tested on against my employer's Exchange 2007 server (which is set-
up to only allow the ActiveSync 12.1 protocol)...

Again, I tried to get the functionality in with minimal impact. The one thing
I'm not 100% pleased with myself is that the ActiveSync 12.1 protocol will
currently always be used when it's available.

A better option might be to allow the end-user the choice of the supported
protcol, but that would have required much more code-impact, so I settled
for this compromise...

For more information, or more discussion, you can drop me an email..

Change-Id: Ifb07791e5094f7a1a719b10d33002d4998f7796e