Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

Hi Keiji,

Thank you for uploading this patch.

Unfortunately, this is a fairly significant change in behavior that we need to think more about because it affects a lot of users that may not appreciate this new pattern.  

That being said, I definitely feel that you are interested in these features, and I will be sure to relay the request to the rest of the team to discuss. We'll prioritize it accordingly and implement it, probably behind a setting, if we decide to do it. 

Thanks, again!

Ben

Patch-set: 1
Reviewer: Ben Komalo <1003964@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1