Update patch set 1
Patch Set 1: Looks good to me, but someone else must approve

Verified by many people, this is the accepted fix if you look at the issues and check the RFC, and this has been merged into CyanogenMod.

Need someone to give this the green light so it can go into AOSP so we can finally close the book on this bug. Thanks.

Patch-set: 1