Update patch set 2
Patch Set 2:

I decided to fix another similar translation issue I found when searching the issue tracker. That's the reason for patchset 2.

The meaning of the Invite tab is to answer the invitation you just received. The current Swedish translation is a translation of the verb invite which doesn't make sense in this context.

Patch-set: 2
Label: Code-Review=0