Update patch set 3
Patch Set 3:

Can you please confirm the SW translation for "Invite" that is to be changed is "Inbjudan"? What's your suggestion?

Thanks Angela

---------------

Hi Angela,

As I describe in the commit message, the original English words used here ("Forward" and "Invite") translate to different Swedish words based on the context in which they are used. In this case these words were translated with the wrong context, leading to a completely different meaning.

If you don't trust a native speaker of the language, you can check the forward translation by translating the sentence "Forward a letter" in google translate. It gets properly translated to "Vidarebefordra ett brev". The same word is used when referring to emails in all email clients.

"Invite" in this context refers to an invitation (ie. an invite), more specifically a calendar invite. The Swedish word used before this patch means "to invite", ie. the verb "invite". Once again, Google Translate gets it right if given the proper context ("Send an invite" translates to "Skicka en inbjudan"). The use of the wrong translation here makes it extra confusing because the meaning is almost the opposite of what the actual feature does.

In summary, the following changes should be made:

"Framåt" should be replaced with "Vidarebefordra". (Not sure what you meant with that browser mention, the use of forward in the browser is correct.)

"Bjud in" should be changed to "Inbjudan".

Acutally, just look at the side-by-side. The word in red on the left should be replaced with the word in green on the right. It's as simple as that. :)

Patch-set: 3