Update patch set 3
Patch Set 3:

Hi Björn,

This is the first time I use this tool, so I am still trying to understand how to display changes side by side. 

Anyway, in the translation console, the only instance of "Framåt" I could find was in the Browser dedicated folder. We won't change this, as it's already correct.
In AndroidEmail, "Forward" is always translated as "Vidarebefordra".
What's the complete English UI string?

Same for "Bjud in". There are no instances in the Calendar dedicated folder in TC. 

Is a screenshot available for these two bugs so I can better check?
Also, do you have access to translation console so I can give you a reference of the strings I'm talking about?

Thanks
Angela

Patch-set: 3