Update patch set 3
Patch Set 3:

If you look just above the Comments section, you'll see a table with the columns File Path, Comments, Size, Diff and Reviewed. In the Diff column, click the Side-by-Side link on the row of the file in question. You could also just click the value in the File Path column since it will default to side-by-side view.

This is the result you will get: https://android-review.googlesource.com/#/c/42682/3/res/values-sv/strings.xml

I don't have a screenshot of the Swedish translation in action as I already have the patched locally like I mentioned before. I included an English screenshot that should explain it though.

https://www.dropbox.com/s/g3zeemvnzxymije/Screenshot_2013-05-29-03-13-12.png

There you see the Invite tab. It used to say "Bjud in" on that tab in Swedish, which is wrong.

I think the translated string belongs to the Email app, not Calendar in TC so that might explain why you didn't find it.

No, I do not have access to the translation console as I'm not a Google employee. TC seems to be an internal system.

Patch-set: 3
Reviewer: Björn Lundén <1005571@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0