Fixed issue where pop3/imap accounts would stop syncing indefinitely.
Because AlarmManager is being called with AlarmManager.ELAPSED_REALTIME_WAKEUP we can not pass in a value of 0 as it is expecting a time based off of SystemClock.elapsedRealtime().
The way it is coded right now, it never executes the PendingIntent for ACTION_CHECK_MAIL when prevSyncTime == 0 or nextSyncTime is in the past.
Therefore you end up with a situation where mail is never checked for pop3/imap accounts.

Change-Id: Id813283fafbf18119724c4cc41e4aba5c9514c24
(cherry picked from commit 3584b902ad834bcb3d090372daa8ba676ca2365d)