Fixed security hole with Exchange implementation.
When a user removes the device administrator the event only triggers a removal 
of account data in the Email app and leaves the account intact. The user can 
then initiate a pull to refresh of their data and synchronize all of their PIM 
data but without a device administrator in place. There is nothing in the sync 
code that verifies a device admin is still in place so if the account is not 
wiped here then the syncing can continue until the Exchange server sends down a 
policy refresh which isn't guaranteed to happen.

Change-Id: I4f11ac7f859d6d7c65a55e47142de13b300482de
Signed-off-by: Daniel Ochoa <gameboy45@gmail.com>