Create change
Patch-set: 1
Change-id: Iade12f56614d0aae48115438d9948235fbf9dbc8
Subject: When synchronizing an Exchange ActiveSync Account with a server which does not sync all "required" mailboxes, local ones will be created with an empty serverId. This will cause the account integrity check to fail as the local mailboxes are recognized as duplicates. This is because of the sql statements not excluding mailboxes with an empty serverId. This will then cause the sychronization to clear all mailboxes and redo an initial sync. By just excluding mailboxes with an empty serverId in the "ACCOUNT_INTEGRITY_SQL" sql statement in the class "EmailProvider", this should be fixed.
Branch: refs/heads/master
Commit: a41d49476d028afa8fad9fa7328f59bc46881fd8