Email: fix crash logging in.
If user switches away from and then return to Email app when logging
into some pre-configured email account, a crash will occur. Logging
is done with AsyncTask. When it succeeds but the fragment is not in
resumed state, result is not delivered until next Fragment onResume()
call. It calls onCheckSettingsComplete() and then proceed(), involving
fragment transactions which is unsafe during FragmentManager state
changes. Delay UI updates to onPostResume().

Change-Id: I4dd370d2fe818db9a04b5093db9756053bbf5560
CRs-Fixed: 2060222